refactor: create Labels in entities
