--- src/Labels.css
@@ -1,29 +1,0 @@
-.labels {
-  margin-left: 1rem;
-  display: flex;
-  gap: 0.25rem;
-  flex-direction: column;
-}
-
-/* .labels__labels {
-  display: flex;
-  align-items: center;
-  cursor: pointer;
-  gap: 0.5rem;
-  padding: 0.5rem;
-}
-
-.labels__labels:hover {
-  background-color: #ececec;
-}
-
-.labels__checkbox {
-  border-radius: 0.25rem;
-  height: 1.25rem;
-  width: 1.25rem;
-}
-
-.labels__text {
-  text-transform: capitalize;
-  font-size: 1rem;
-} */
