+++ src/Sidebar.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import "./Sidebar.css";
 import CreateEventButton from './CreateEventButton';
 import SmallCalendar from './SmallCalendar';
-import Labels from './Labels';
+import Labels from './entities/labels';
 
 const Sidebar = () => {
   return (
