+++ src/entities/labels/style.css
@@ -1,0 +1,6 @@
+.labels {
+  margin-left: 1rem;
+  display: flex;
+  gap: 0.25rem;
+  flex-direction: column;
+}
