+++ src/shared/ui/index.js
@@ -1,0 +1,1 @@
+export { Label } from './label';
